+++ geolocator_platform_interface/test/src/enums/location_accuracy_test.dart
@@ -2,10 +2,10 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:geolocator_platform_interface/geolocator_platform_interface.dart';
 
 void main() {
-  test('LocationAccuracy should contain 6 options', () {
+  test('LocationAccuracy should contain 7 options', () {
     final values = LocationAccuracy.values;
 
-    expect(values.length, 6);
+    expect(values.length, 7);
   });
 
   test("LocationAccuracy enum should have items in correct index", () {
@@ -17,5 +17,6 @@ void main() {
     expect(values[3], LocationAccuracy.high);
     expect(values[4], LocationAccuracy.best);
     expect(values[5], LocationAccuracy.bestForNavigation);
+    expect(values[6], LocationAccuracy.reduced);
   });
 }